# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "info@preprlabs.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth API - Login"
                ],
                "summary": "Send request for check login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter email of related to account!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter password of account!",
                        "required": true,
                        "explode": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found!"
                    },
                    "400": {
                        "description": "Bad request!"
                    },
                    "500": {
                        "description": "Internal Server Error!"
                    }
                }
            }
        },
        "/api/v1/auth/verify-two-factor": {
            "post": {
                "tags": [
                    "Auth API - Verify Two Factor"
                ],
                "summary": "Send request for check Verify Two Factor",
                "operationId": "verifytwofactor",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter email of related to account!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "description": "Enter two factor otp!",
                        "required": true,
                        "explode": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found!"
                    },
                    "400": {
                        "description": "Bad request!"
                    },
                    "500": {
                        "description": "Internal Server Error!"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth API - Check Register"
                ],
                "summary": "Send request for check register",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "name": "preferred_language",
                        "in": "query",
                        "description": "Enter langauge id for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Enter username for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter email for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Enter first name for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Enter last name for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter password for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "confirm password same as password!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "device_platform",
                        "in": "query",
                        "description": "Enter device platform for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "user_type",
                        "in": "query",
                        "description": "Enter user type for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "purpose",
                        "in": "query",
                        "description": "Enter purpose for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "country_code",
                        "in": "query",
                        "description": "Enter the country code for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "Enter the phone number for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "two_factor_verification",
                        "in": "query",
                        "description": "Enter the two factor verification for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "organization_name",
                        "in": "query",
                        "description": "Enter organization name for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "vanity_link",
                        "in": "query",
                        "description": "Enter the vanity link for registered!",
                        "required": true,
                        "explode": true
                    },
                    {
                        "name": "referal_code",
                        "in": "query",
                        "description": "Enter the referal code for registered!",
                        "required": true,
                        "explode": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found!"
                    },
                    "400": {
                        "description": "Bad request!"
                    },
                    "500": {
                        "description": "Internal Server Error!"
                    }
                }
            }
        },
        "/api/v1/auth/forget-password": {
            "post": {
                "tags": [
                    "Auth API - Forget Password"
                ],
                "summary": "send email for forget-password",
                "description": "send otp for forget-password",
                "operationId": "forgetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email uses for send otp",
                        "required": true,
                        "explode": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found!"
                    },
                    "400": {
                        "description": "Bad request!"
                    },
                    "500": {
                        "description": "Internal Server Error!"
                    }
                }
            }
        },
...
        "/api/v1/auth/send-otp": {
            "post": {
                "tags": [
                    "Auth API - Send Otp"
                ],
                "summary": "Send request for send otp",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Uses emails to get account!",
                        "required": true,
                        "explode": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found!"
                    },
                    "400": {
                        "description": "Bad request!"
                    },
                    "500": {
                        "description": "Internal Server Error!"
                    }
                }
            }
        },
...
    }
}